[compile] Stop unconditionally patching constrain_to_fx_strides (#36152)
